Trading Bot/requirements.txt.txt
python-binance==1.0.17
requests
tk